1. Hilt refactoring with clean architecture

+++ app/src/main/java/com/example/firechat/chat/presentation/adapters/HalfwayLocViewHolder.kt
@@ -31,7 +31,7 @@ class HalfwayLocViewHolder(
         val currentUserId: String = Constants.currentUserId
         val fromUserId = message.messageFrom
         val sfd = SimpleDateFormat("dd-MM-yyyy HH:mm")
-        val dateTime = sfd.format(Date(message.messageTime))
+        val dateTime = sfd.format(Date(message.messageTime as Long))
         val splitString = dateTime.split(" ").toTypedArray()
         val messageTime = splitString[1]
