+++ app/src/main/java/com/example/firechat/chat/presentation/adapters/LocationRequestViewHolder.kt
@@ -33,7 +33,7 @@ class LocationRequestViewHolder(
         val currentUserId: String = Constants.currentUserId
         val fromUserId = message.messageFrom
         val sfd = SimpleDateFormat("dd-MM-yyyy HH:mm")
-        val dateTime = sfd.format(Date(message.messageTime))
+        val dateTime = sfd.format(Date(message.messageTime as Long))
         val splitString = dateTime.split(" ").toTypedArray()
         val messageTime = splitString[1]
